app/requirements.txt
Flask==2.1.0
google-cloud-firestore==2.11.1
google-cloud-secret-manager
google-cloud-storage==2.10.0
gunicorn==20.1.0